Download functionality for Spectra page
Adds download functionality to the Spectra page, utilizing the library JsZip.

Sadly, this breaks the expand/hide functionality that was implemented on this page before. This will need fixing in the future, should the expand/hide functionality still be desired.

This all still requires testing when downloading a large Number of files.